pallet_detection/requirements.txt
torch
torchvision
ultralytics
opencv-python
pycocotools
Pillow
